split localmem_naming_tests from server_tests
Summary:
It was kind of lazy of me to include sIDE-tests inside server_tests before. I did it because that was a convenient place that set up files and naming table.

But I'm going to want to write more localmem-naming-tests in future, so it's time to split it up now.

This diff is mostly just trivial splitting. The only non-trivial thing is that I changed these naming tests from folded-decls to shallow-decls.

Reviewed By: arxanas

Differential Revision: D21366719

fbshipit-source-id: 31270a1755aa12f1b771397cc624cddf9fc16166