Future: handle exceptions from continuations
Summary:
Discovered this while writing code that relies on `is_ready` checks and that was intentionally throwing exceptions in a continuation.

We already correctly handle this inside of `Future.get`, so this is diff just applies the same handling to the continuation inside of the `is_ready` function (also called by `check_status`).

Reviewed By: ljw1004

Differential Revision: D21489408

fbshipit-source-id: 1f4722ed35ec3fc368fd63b3c65b4049419f3da4