Informant: don't redownload previously-downloaded saved states
Summary:
As I was testing the rest of this diff stack, I observed that the informant always fetches saved states from Everstore, even if they are already in the cache.

In State_loader, we already check for cached states that don't directly correspond to the merge base.
This diff is just the same change, but in Informant.

Reviewed By: ljw1004

Differential Revision: D21489604

fbshipit-source-id: 8f348599d860d182d8aa894fd45f510ac8204c92