Informant: ensure state fetching logs go the into monitor log
Summary:
The informant used to need to prefetch saved states in a separate process.
The downside of this is that the process doesn't log into the monitor log, and it isn't able to log into our Scuba tables, either.

This diff removes the extra hop, since I previously made State_loader functions async.

Reviewed By: ljw1004

Differential Revision: D21489406

fbshipit-source-id: ce0f5cafa7e8fcdef1b137be8a613ea4f96c3c7b