Provide functions which allocate unescaped strings in an arena
Summary:
Since string literals can occur in decls (in global const declarations and in attributes), the direct decl parser needs the ability to unescape them in order to interpret them correctly. For performance reasons, we are avoiding using the global allocator and copying strings as much as possible. This makes the current API of the escaper module a problem--it always allocates a new String for the unescaped content. Since the output of the decl-parser is arena-allocated, we cannot (easily) refer to a String without leaking memory, so we would need to copy the contents of the String into the arena.

To avoid the String allocation and the copy into the arena, this diff abstracts over std::vec::Vec and bumpalo::collections::Vec in the unescaping functions. To avoid further copying when removing quotes around a string literal, this diff revises `extract_unquoted_string_` to return a subslice of the input (and renames it to `unquote_slice`).

This diff also fixes a safety hole in `extract_unquoted_string`, where `start` and `len` could have specified a subslice that did not begin and end at valid UTF-8 codepoint boundaries. It now returns an error instead of an invalid String.

I removed two test cases and revised two test cases dealing with unquoting heredoc strings like `"<<<EOT\nEOT;"` and `"<<<EOT\nEOT;\n"` because the lexer does not include the trailing semicolon or newline in a heredoc string literal. They are understood by the lexer as a normal semicolon token and its trailing trivia.

Reviewed By: shiqicao

Differential Revision: D21462694

fbshipit-source-id: e8864bd6bfca62f200f9d577b7ed82b30404f2ef