Enable typehint notices
Summary:
HAC typehint notices are now flat. Enable them by default. We need to make the remaining tests HAM-safe to land this change.

The reason to land this change to master rather than turning it on via config is that without updating FDO + BOLT, it looks like typehint notices are still a regression in Swayze, probably because they activate code paths that are currently dead.

Reviewed By: ricklavoie

Differential Revision: D21460733

fbshipit-source-id: 80970294ed1bdeb36c841fcea1751e61c7a739cd