Implement a Stateful NAST Check
Summary:
Currently, the nast_check visitor is a bit unwieldy to use for stateful/context based error checks. Each method of a nast check visitor is context free but there are many error checks that require some context/state to correctly report errors. Currently, many of these checks do one of the following to gain context/stateful capabilities.

1. Add context/state to the global nast check env
2. Retain context/state locally in the visitor via a stateful reference
3. Implement another stateful visitor and call into that

There are a couple of problems with each:
1. The code to do this is a non obvious pattern. It also pollutes the global environment with state that is only needed in one particular visitor
2. Updating a reference statefully is difficult to reason about within the visitor framework
3. This breaks out assumption that we will only make one pass over the ast for all checks, instead of one per check.

This diff creates a framework for a stateful nast check visitor where each visitor is able to retain its own state when iterating over the AST.

Reviewed By: Wilfred

Differential Revision: D21292538

fbshipit-source-id: 9fe6929435fed28a177c7cc2fe82c896a90c97a9