Do not report duplicate cyclic definition errors due to expanding type aliases and type constants
Summary:
Reporting of cyclic definitions (through type aliases, and through type constants on classes) currently produces duplicated errors. Worse, errors will be reported when checking a file that just *references* a type that is involved in a cycle, even though that file has no error. For example:
```
////file1.php
<?hh
function foo(D::T $x):void { }

////file2.php
<?hh
class D {
  const type T = D::T;
}
```
When checking `file1.php`, Hack will report an error, located in `file2.php`.

It should be enough just to report an error in the *definition* of a type involved in a cycle, not in the *use* of that type. That's what this diff achieves, by carrying around an extra field `report_cycle` in `ety_env` that specifies the name of a type whose definition (by `type` alias or `const type`) is being checked. Cycles are still always detected, but only reported if the type in the cycle matches the contents of the `report_cycle` field.

Also: I have removed the `HH_FLAGS` field from the `tconst` test directory in order to see *all* of the errors that are produced from Hack. Some were getting hidden.

Reviewed By: manzyuk

Differential Revision: D21510420

fbshipit-source-id: 6bc1cb43397979803b75beb7a4ca3e6a54cb2b8b