Fix error for const overriding so that overriding property is primary position
Summary:
The following code
```
<?hh // strict

class A {
  <<__Const>>
  public int $x = 4;
}

class B extends A {
  public int $x = 42;
}
```
produces this error:
```
File "subclass_prop_not_const.php", line 5, characters 10-12:
This property is __Const (Typing[4204])
File "subclass_prop_not_const.php", line 9, characters 10-12:
This property is not __Const
```
Unfortunately, the "primary error position" (used for HH_FIXMEs etc) is in the overridden class (A) not in the overriding class (B) that is the actual point where the error is detected.

It should be the other way round, as with similar errors involving overriding.

Differential Revision: D21511727

fbshipit-source-id: 2ba8d3ebc4f3111caeffd7e4e96fe463b4257c2b