Add cursor system
Summary:
`hh_fanout` is intended to be a short-lived process. To reuse computation between invocations, it writes its current state to disk and reads it on later startups.

This commit is a reference implementation that is backed by an OCaml blob which is marshalled to disk. (This will almost certainly be too slow for production, but it's much faster to prototype. The intended final version will store its state in a SQLite database.)

It would be dangerous to mutate state in-place, since if the caller of `hh_fanout` requested a mutation and then crashed, it may start up and incorrectly re-request the same mutation, leading to an inconsistent state. Similarly, if the the caller of `hh_fanout` requests a mutation and `hh_fanout` itself crashes, the caller doesn't know whether to retry the operation.

The solution is to implement `hh_fanout`'s state on disk as a persistent data structure (in the technical sense of a data structure for which old versions are kept). Queries that may change `hh_fanout`'s state return a "cursor", which points to the resulting state. If the caller crashes, then it is safe to re-request the same computation from the previous cursor again.

Reviewed By: ljw1004

Differential Revision: D21415417

fbshipit-source-id: 83bc648d2f5c403b40755927e4c9ebb99e6373d5