workers shared HackEventLogger globals
Summary:
The overall goal of these three diffs is to be able to send HackEventLogger data for typechecker bugs, but without overwhelming HackEventLogger in case some bugs are just too frequent.

I want a way for HackEventLogger state to be shared between master and workers. That's what this diff does.

It would have been possible to split it into four methods rather than just two
* serialize_from_master_to_workers
* deserialize_from_master_to_workers
* serialize_from_workers_to_master
* deserialize_from_workers_to_master

But the current two methods are equally expressive and (I think) easier to follow.

Reviewed By: CatherineGasnier

Differential Revision: D21481677

fbshipit-source-id: 02e57f04f8e097c5d6b2647521f860b4a2100881