Split error codes for read/write on array access and object access
Summary: This diff splits the codes for read and write on certain array access and object access expressions so that the former can be tracked and eliminated. The motivation is that when there's a error on a read that is fixme'd, it generates a Tany (Terr) that flows into other types in a function body.

Reviewed By: Wilfred

Differential Revision: D21421657

fbshipit-source-id: ed89bc6cd81c124f40b84c269109ca76f4e56290