Add RFunc to the `DataType` enum
Summary:
This diff adds RFunc to the `DataType` enum. This forces alot of `switch`/`case` statements to need to be updated with `KindOfRFunc`.

Some of these I have no idea what to do - so I've left them as errors/TODOs. Others I have attempted to fill out with behavior that I think is correct.

Reviewed By: paulbiss

Differential Revision: D19745777

fbshipit-source-id: 441df2a08c6c19689c946457620662c2abcc81bc