Move towards explicitly whitelisted arena-compatible types
Summary:
In order to prevent future memory-leaks, the `ArenaCompatible` trait is added, to allow us to construct a "whitelist" of arena-compatible types.

This diff implements the `ArenaCompatible` trait for custom types.

It also tries to hide the underlying arena implementation (bumpalo), because using bumpalo directly can be an other source of memory leaks. In the future we might want to fully abstract away the actual arena implementation. Currently this is difficult because we depend on the bumpalo-provideded vector type.

Differential Revision: D21502202

fbshipit-source-id: 13e393256abd991b219515a96809d6668b5f06fc