Update the jit.tcspace.*.used counters when we emit code
Summary:
Right now the jit.tcspace.*.used counters are only updated as a
side-effect of calling reportJitMaturity(). This has at least two issues:

1) Once we've decided the JIT is "mature", we'll never get far enough
into the function to update the counters. As a result, the counters
stop updating, even if we're still adding more code. This is probably
not the behavior people expect from these counters.

2) reportJitMaturity() is only called after retranslate-all, and when
a request exits. We'll essentially defer all of the code a request
generates until it ends. This can be a problem for a long running
request, as we use the rate of code generation from these counters as
a sign that we're warmed up.

Move this updating logic out of reportJitMaturity() and into its own
function updateCodeSizeCounters(). Call this after we're done emitting
a translation into the TC. So, the counters will accurately reflect
the current size of the TC, and the rate of code generation will be
representated more accurately as it won't be deferred until request
end.

Reviewed By: ottoni

Differential Revision: D21541958

fbshipit-source-id: 0a3c83c3459b109f942f506cc0eaec7d7b7d93dc