better hg diff, status
Summary:
This fixes up two issues with hg rage:

(1) it only did "hg diff", and so didn't pick up modified files that weren't yet tracked by source control. This diff adds a call to "hg status" to show them.

(2) it had a bug where it wrote the wrong contents into the "www hgdiff.txt" attachment. It was writing the output of clowder, not the output of "hg diff". I've fixed this.

Reviewed By: arxanas

Differential Revision: D21559844

fbshipit-source-id: 7ecc63b49f511c465ecb25668b30e91c52f5faa7