Stop doing useless extra work to display serial total time.
Summary: We already log the time for every pass/fail/skip result in the 'time' field, which already gets returned to the parent via the test results json. Stop also writing these times to 'trun' files and then scraping the 'trun' files to compute the total serial execution time. Walk the results instead (which we already have in memory) to get the exact same data.

Reviewed By: fredemmott

Differential Revision: D21518815

fbshipit-source-id: 38dbd9b4a52e6e585ac31033bfc334547b4dcbf6