Fix to_varray tech debt in run.php
Summary:
In run.php we passed the HHVM commands around as either strings or varrays of strings (depending on if we had 1 mode or 2). This was super-confusing, had bugs, and required users that definitely wanted a varray to jump through hoops (the aforementioned to_varray helper).

Always produce and return varray<string>, even when we only have 1 mode. Add types to a number of function signatures to make sure we're doing this right. Rename run_one_config to run_foreach_config to reflect what it does. Reject use of modes together with `--hhbbc2`, since it doesn't run code (only compares hhas).

While we're here, move adding timeout_prefix into run_config_cli so that all the callers don't have to do it.

Reviewed By: oulgen

Differential Revision: D21513500

fbshipit-source-id: 26e607af6d44bf99cb29e25db77534c08feba2a9