Remove run.php support for `-m pgo`
Summary:
`run.php -m pgo` sets `-vEval.JitPGORegionSelector=hotcfg` but this is the default (which --wholecfg overrides), and sets `-vEval.JitPGO=1` (which is the default unless you define HHVM_NO_DEFAULT_PGO when you build). We otherwise enable PGO with flags like `--retranslate-all` and `--jit-serialize` (together with other settings), not a mode.

Remove support for `-m pgo`, as we rarely build with HHVM_NO_DEFAULT_PGO outside of contbuilds. This leaves interp and jit as the only valid modes.

Reviewed By: oulgen

Differential Revision: D21513502

fbshipit-source-id: a9b8ae69d9ddf098c625727aec25015b8c565972