Require attributes arguments to be literals in lowerer
Summary:
Attribute arguments that are not literals pass the typechecker, but cause HHVM to fatal. We want to warn users about invalid attributes during the parsing pass so that they will not be able to ship code to production that will fail.

The word "literal" here is used pretty loosely and includes:
* Regular Literals (eg string, int, etc)
* Collections of literals (eg vec)
* Concatenated strings
* `::class` strings
* shapes

HHVM seems to support numeric arithmetic too, but I've decided not to allow it.

Reviewed By: Wilfred

Differential Revision: D21526490

fbshipit-source-id: 3e91fd0e2e2014aad5db96b462839bc3727a03e9