Discriminate between final and deferred type checking
Summary:
It's confusing to look at the logs when defer_class_declarations_threshold is set because we attempt to type check files more than once.
This is just changing the tag in the server log that we use for when we finished type checking and when we just attempted to type check a file but only discovered its decl dependencies instead.

Reviewed By: ljw1004

Differential Revision: D21589341

fbshipit-source-id: 1f062ccbc81ab5a8d7bad96516bc11fe757ada14