Add hackpartial to runtime
Summary:
This fixes hphp/test/run and a few other spots to pick up `.hackpartial` files.

Also, there was a bug where a `.hack` and `.hackpartial` file with the same contents would give the same test output, even though parse errors could be different. I added a char to `mangleUnitSha1` to mitigate this.

Reviewed By: fredemmott, markw65

Differential Revision: D21542278

fbshipit-source-id: 4bd9537dd1c2c91a1abecd8faeade5e5d627ddde