Eliminate live relocation
Summary:
When the EvalPerfRelocate is set, then after that many requests, we'll relocate functions in the TC. I suppose the benefit of doing this relocation is that we can test that we perform this "live relocation" correctly and invalidate all translation metadata that goes stale when we do so, but since we don't actually use this mode in production, it's not useful to maintain it...

markw65 notes that relocation may have been useful for ARM since indirect jumps are quite expensive there, but now that we smash all the jumps in optimized translations in JIT deserialize mode, that benefit is also greatly reduced.

Delete this option, the reloc debug map file, and the live relocation logic.

Reviewed By: markw65

Differential Revision: D21610436

fbshipit-source-id: ffb60afff9ba2d936ff482d7541aef440efd51e3