Add options to handle cold translations
Summary:
See D21610978 for rationale. The next step is to add options that will identify and lay out cold translations correctly. Right now, "hot", "cold" and "frozen" thresholds are relative to a translation's entry, which won't suffice for this purpose. We need an alternate way to identify what a hot translation is. Here are a few options:
1. A fixed absolute weight cutoff.
2. Some kind of global hint, incorporating profile data from all translations.
3. A per-function hint, incorporating profile data across translations for a single function.

Of these, I think approach 3 is best; we're already sorting the TC by function weight, not by translation weight, so if there are multiple translations and one's much hotter than the other, then we probably want to put the cold one into a different area. This situation comes up with polymorphic functions (e.g. that usually take a dict but that may get a Map object).

In this diff, we implement approaches 1 and 3. We also implement a few other small options which help with other patterns we observed. The full list of new options is:
- **JitLayoutMin(Hot|Cold)Threshold:** (approach 1) absolute thresholds. Small values (e.g. 100 and 1) may be useful even if our main hint is the "hot weight".
- **JitPGOVasmBlockCountersUseHotWeight:** (approach 3) in region selection, collect the entry weight of all translations for a function and pass that as a hint to vasm-layout.
- **JitPGOVasmBlockCountersSkipFixWeights:** weights may appear to be incorrect because of hidden edges in fallbackcc / bindcc / etcs. ops. Don't "fix" these weights.
- **JitLayoutSeparateZeroWeightBlocks:** don't include zero and non-zero weight blocks in the same cluster.

Reviewed By: mofarrell

Differential Revision: D21610987

fbshipit-source-id: 0d8147cc12eba59fde0caca0333745bb26040c14