Allow entries in arbitrary areas
Summary:
ricklavoie and I have been examining the production behavior of vasm-block-layout. We found what we believe is a significant performance opportunity: the algorithm currently puts a lot of cold code, including side exits and catch traces, into Main. In particular, as soon as a function is too cold, the cutoffs for main will drop to 0 and the *entire translation* will be put in Main. Decreasing weights can actually increase the amount of code in Main!

To fix this issue, the first step is to add a mechanism for allowing the JIT to generate translations where the entry is in cold or in frozen. In this diff, we make such code possible.

We add a "TransLoc::entry" method, which returns the code address of the first non-empty area for a given translation. As long as we ensure that no block is in a hotter area than the entry block, this address will be the correct entry. This restriction is reasonable, because it isn't a good idea to generate code where some hot internal block is in Main but the entry is in Cold. I don't care to relax it further right now.

As for the rest of this diff, we simply audit all callers of mainStart and see whether they really mean mainStart or if they should use entry instead.

The new mechanism doesn't work if we pad TC entries (which we intend to do when the TC is reusable). I noticed that we're padding these entries in production, too, which isn't intended. I'll see if we can get any benefit by setting that to 0 along with the other options introduced in the next diff.

Reviewed By: mofarrell

Differential Revision: D21610978

fbshipit-source-id: f98fbc582c2617d1c93fce16d1637ca2a530ddae