Kill Eval.DisableSomeRepoAuthNotices
Summary: This flag appears to only control whether we raise these particular notices for specialized property lookups in repo mode (so it's both inconsistent between repo and non-repo mode and inconsistent within repo mode). The prevailing theory is that we added the flag when these notices were first added because/in case they proved to be regressions in production. That doesn't appear to be the case anymore so remove them.

Reviewed By: oulgen, jano

Differential Revision: D21598560

fbshipit-source-id: 5c5fb05c71adcbf64e9c4070a735a147c1080b33