monitor uses sync scuba
Summary:
We've had a long-running mystery where the monitor doesn't respond to clients. We've made attempts to solve it but never managed to pin it down.

Just recently we got some surprising behavior which may be a clue:
1. We observed that serverMonitor wasn't calling "flush" during its normmal operations; it sent a few messages to the scuba daemon around startup, and then only buffered them up until the monitor process finally terminated.
2. The commbination of EventLogger.init_fake and EventLogger.flush together in the monitor caused failures - the attempt to send data to the scuba daemon got an EINVAL in its select call, and the client<->monitor handshake which uses Marshal...preamble to exchange a string version number was failing in the Marshal step. We'd observe the client making countless requests to the monitor, but they only rarely succeeded.
3. Removing EventLogger.flush allowed it to succeed
4. Changing it from EventLogger.init_fake to using real scuba logging allowed it to succeed

My suspicion is that all these things are related. I believe there's some bug in the way the monitor speaks to the scuba daemon which results in it getting mis-wired FDs. This might be because of of duping wrong FDs when we fork, it or might be because of code (which I don't understand) which picks FDs for the scuba daemon, based on initid, which is actually SHARED I think between monitor and server:
https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/facebook/logging/hackEventLogger.ml?commit=6589dffe1f92dbe6cbc67b1637b89ded1f117f8a&lines=427-436

I think we have a safe clean opportunity to reduce potential exposure to FD bugs. We can simply make the monitor use synchronous scuba logging. The monitor doesn't write many logs, so this won't impede perf. And it'll remove entirely the possibility of monitor FD confusion. That's what this diff does.

What I expect this diff will do: (1) it will restore logging to the monitor process, which prior to this diff only logged a handful of events at startup and then the rest upon termination; (2) it might reduce our exposure to the unreproducible+untestable monitor connection issues.

Note: from earlier events, I observed that synchronous logging to scuba is no slower than using the daemon. I presume the reason is that invoking `ptail` incurs only a tiny startup cost, and then ptail does the minimum work to queue up the log, and then returns. Which is basically exactly what our scuba daemons do.

Reviewed By: arxanas

Differential Revision: D21613712

fbshipit-source-id: cadbac1d69b2ad9a682f710a28b21ca9c46244b7