Track set of profile TransIDs and reduce TargetProfile counters across them
Summary:
Due to guard relaxation, a region block in an optimized translation can actually correspond to multiple profile translations that were merged together. When that happened, the block and instructions in the optimized code was only associated with one of those profile translations, and the TargetProfiles were only reduced for that one translation.

This diff keeps track of all the profile translations that are merged together by guard relaxation. The single TransID that was tracked in TransContext, IRGS and BCMarker is replaced by a set of TransIDs. For optimized translations, the TargetProfiles can now have a list of RDS keys and links, which are all reduced together.

Differential Revision: D21386699

fbshipit-source-id: f0a638dae8fdeb33274922755e9c94cf1bc15124