Remove HintValue
Summary: The last remaining variant of HintValue, HintValue::Access, does actually require some special handling. Type constant accesses are nested in the CST. If we have `this::Foo::Bar`, the CST looks like `TypeConstant(TypeConstant("this", "::", "Foo"), "::", "Bar")`, but the decl looks like `Taccess(Tthis, ["Foo", "Bar"])`. We need to flatten the nested type-constant-accesses into a list of type constant names. While we could immediately construct types in `make_type_constant`, it would be wasteful--we'd construct `Taccess(Tthis, ["Foo"])`, then need to copy the entire type (including the entire list of typeconst names) for each containing typeconst access node. Instead, we can allocate a Vec for the innermost node, and mutate it as we build the outer nodes. This diff migrates this handling from HintValue to a new Node_ variant, Node_::TypeconstAccess.

Reviewed By: Wilfred

Differential Revision: D21218097

fbshipit-source-id: 628bc0bc861cd81e36739de30677252ab676d838