Handle illegal declarations outside namespace bodies
Summary:
If a file contains a namespace declaration with a body, the file must contain no code outside a namespace body.

We used to have a test case ensuring that we still produce a decl (equivalent to that produced by OCaml) for files containing invalid syntax of this kind, but I broke the behavior of the decl parser (and deleted the test case) for this invalid syntax in {D20840140}. This diff produces a sensible decl instead of panicking, and reintroduces test cases.

Reviewed By: Wilfred

Differential Revision: D21342032

fbshipit-source-id: 9e05c0711d92abb6213af4d3cb78725464c0e8bc