Remove IR for method and property decls
Summary:
As of {D21192276}, when we encounter a method header introducing a type parameter, we push it onto a stack of type parameter names, so that we can track which type parameters are in scope. When we invoke `function_into_ty`, we pop the type params stack, since presumably we have just exited the method and are allocating the function type of its decl.

Only, that isn't quite how it works--when we exit the method and invoke `make_methodish_declaration`, we currently create a `FunctionDecl` rather than the actual function type. These `FunctionDecl`s are aggregated in `make_classish_declaration`, and only *there* do we invoke `function_into_ty` for each. This is because historically, we needed to feed the type parameters introduced by the class declaration into `function_into_ty`, so we needed the `FunctionDecl` IR to defer the construction of types until `make_classish_declaration`. The result is that our pushes and pops to the type parameters stack are not paired correctly. In the added test case, we push Ta, push Tb, create the function type for f, pop Tb, create the function type for g, and finally pop Ta. This results in the Tb in g's signature not being correctly recognized as a type parameter (and a type Ta in g's signature would incorrectly be recognized as a type parameter).

Instead, we should construct the function type in `make_methodish_declaration`. This diff does so. Similarly, this diff constructs property decls in `make_property_declaration` instead of using a `PropertyDecl` IR.

Reviewed By: Wilfred

Differential Revision: D21309885

fbshipit-source-id: 71d6760e8126493403c1a3c14d0bbb53d3645141