Remove OperatorType
Summary: This type appears to be a subset of TokenKind (with some of the variants renamed to their ast_defs::Bop equivalents). Since we never match against it exhaustively, I think it'd be simpler to just use TokenKind. I removed OperatorType::Xor and OperatorType::LogXor because they are currently unused.

Reviewed By: Wilfred

Differential Revision: D21337572

fbshipit-source-id: 7fc37bb5ec036e2ccb6a15e30d68a7a621afaf69