Modify serverBigCode to parse all tokens before the target token
Summary:
Previously we only extract all identifier tokens before the target as our context. However, this is no longer used in production for ranked autocomplete, and we need to extract all tokens including syntactic tokens and punctuations.

In this diff, we use the `tast` for identifying the target tokens (member access, local var), and use the full fidelity syntax tree to extract the context tokens.

Reviewed By: skim9

Differential Revision: D21617954

fbshipit-source-id: 1dafec236ee7827f77e9c86cc64000cbf9b2c2ba