Fix thread to LightProcess Mapping
Summary:
It was using the bottom bits of pthread_self to choose a
LightProcess. Unfortunately, on linux, pthread_self is the address of
the thread local section, which is a fixed offset into a page aligned
data structure; so the bottom 12 bits are always the same, and in
practice, we only ever used one LightProcess.

Lets use hash_int64 to mix it up a bit.

Reviewed By: oulgen, paulbiss

Differential Revision: D21649350

fbshipit-source-id: 149d836e2958af40ee091d0b900c39a075c2721c