richer failure information in clientConnect
Summary:
I'm trying to investigate why people are getting connection failures.

I already know on the monitor side that it happens when monitor receives a version string over `client_fd`, and tries to write back "Connection_ok", but it thinks the fd has been closed by the receiver (gets EPIPE "write").

Why would it think that? What has client been doing? I wish to add more logging to the client's attempt at connecting.

This diff doesn't yet add more logging. What it does instead is just propagate the information that lead to the failure - either an exception arose or a timeout happened. In both cases this information gets bubbled up to where the client is going to log it (in the next diff)

Reviewed By: arxanas

Differential Revision: D21652596

fbshipit-source-id: a3da5f3cc2ebd3cb5c5612eab5b7aabbd3331b7e