Log when monitor gets SIGPIPE
Summary:
There have been steady user complaints about "monitor connection failed". I've been gradually adding more logging around it.

The recent round of logging and rage reports revealed that the monitor does Marshal.from_fd_with_preamble to receive a version string from the client over `client_fd` as expected, and then the immediate next thing it does Marshal.to_fd_with_preamble to send `Connection_ok` back. This writeback is getting a EPIPE "write" error. According to the docs, this means that the FD has been closed on the receiving end; it is accompanied by the process receiving a SIGPIPE.

We already ignore SIGPIPE. All this diff does is add some logging, so we'll get a timestamp when the SIGPIPE occurred. I'm not especially hopeful. But maybe it will shed some light. I've thrown in a callstack as well, not that I'm sure it will be correct, but it at least won't harm. And who knows? it might be valuable?

Reviewed By: arxanas

Differential Revision: D21642330

fbshipit-source-id: d6008aed07c21e93390f9142e2a6a605418c715b