Add Future.on_error
Summary:
We only offer `continue_and_map_err` as a continuation on an existing future, but that function produces a future of a different type.

Adding a convenience function that allows attaching a function that will be executed only in case of error.

The use cases include logging and cleanup.

In the past, in order to execute cleanup steps, I've resorted doing this:
- calling `continue_and_map_err` on future A to produce a future B
- calling `continue_with` on B and matching on result
   - on success, return the value (not wrapped in Result)
   - on failure, throw an exception, wrapping the Future's error in an exception

The downside of this kind of error handling is that it destroys the original Future.error and turns it into a string. The caller then doesn't have the option of examining the error's fields, and is forced to deal with the error message that is either too verbose or not verbose enough.

Reviewed By: ljw1004

Differential Revision: D21633358

fbshipit-source-id: 8e82ec1348376d118b7be98ee4049527e5af80e0