Make order of aliased namespaces canonical
Summary:
The order of array elements (string pairs) is arbitrary in JSON array associated with key

  hhvm.aliased_namespaces

which makes comparison needed for migrating the logic to Rust hard.

Before a more drastic refactoring (.e.g., using a map instead of a list),
let's make sure that having a well-defined order solves the inconsistencies.

Reviewed By: shiqicao

Differential Revision: D21672177

fbshipit-source-id: ec2ccd32dc247e2b4b151c37514f2eac0712d221