Loosen staticness when indexing into non-static ArrLikeVal
Summary:
It is possible to infer the exact value of an array that is non-static at
runtime, which is represented by `ArrLike` type with `ArrLikeVal` tag. When
we index into such value, there is no guarantee that the result is static,
but a bug caused us to infer the result of such indexing as static.

When this happens, hhbbc emits incorrect `AssertRATStk` that causes a failed
assertion at runtime if compiled in debug mode.

Fix this by loosening staticness if the original value was not static.

Reviewed By: markw65

Differential Revision: D21635301

fbshipit-source-id: 980dfa28191554297832631d0388c13776ab8756