Support logging HackC options in HHVM (via JSON)
Summary:
Support logging a list of JSON strings from HHVM, so that `hh_single_compile` can be safely migrated to Rust.
Read the "config_jsons" field from a compile/parse/facts request, and log it as array of strings using `StructuredLog`, *unless* a non-array is passed in which case it isn't logged.

This is mostly to ensure that HackC options are parsed consistently between Rust and OCaml.

**Note**: `hh_single_compile` code should be careful not to pass an actual array (which triggers logging) on *every* compile/parse request, so that the quota is not exceeded.

Reviewed By: alexeyt

Differential Revision: D21483465

fbshipit-source-id: fc9bb5e37c7f4219908f937cf597d1cae0eb1618