Check metadata while backtracing
Summary: Production machines are coring with an invalid lookup of a metadata local. We're not sure whether it's because the local has already been freed, or because we're looking for it in the wrong place. This assertion will cause us to core while backtracing and should preserve sufficient information for debugging.

Reviewed By: jano

Differential Revision: D21698439

fbshipit-source-id: 5ef431538a992a9959e509ba6868da516a146a03