Have array_shuffle return darrays
Summary:
With D21690810, array_shuffle preserves varrays but downgrades darrays to plain arrays. Let's just have it convert return darrays for both the darray and plain array case.

The reason I didn't initially do that is because of how array_shuffle is typed, but I think it's okay - the Hack typing is array_shuffle<Tv>(inout Container<Tv> $array), which this change still respects, so we're not introducing further type-checker unsoundness.

Reviewed By: dneiter

Differential Revision: D21703928

fbshipit-source-id: b7293a157eb86a6dd78c3b00ed0218b0c571cf6e