Replace EmptyArray with PlainArray
Summary:
On the continuing quest to eliminate EmptyArray and plain PackedArrays...

In this diff, we get rid of EmptyArray - if anyone asks for an empty plain array, they'll get a static empty MixedArray.

We'd like to add similar invariants checks for MixedArray that we have for PackedArray, but the problem is that we actually do have plain MixedArrays. To make our checks strict, we introduce a new ArrayKind: PlainArray. This ArrayKind corresponds to the type of a plain array and the layout of a MixedArray.

Creating a new ArrayKind seems like it's introducing complexity. In fact, it's a simplification. The new ArrayKind doesn't require new logic, because it's backed by an existing layout. Adding it gives us the following strong invariant: **an value is a varray if and only if it's TPackedArr, and it's a darray if and only if it's a TMixedArr**. That means that we can actually eliminate the dvarray specialization from the JIT type system. We can even eliminate the dvarray bits altogether! I'll do those pieces next.

Introducing a new kind here also simplifies our work on bespoke array-like layouts. When bespokes enter the picture, our invariant becomes the following: an array is a vanilla varray if and only if it's a TPackedArr, and a vanilla darray if and only if it's a TMixedArr. That will let us optimize vanilla type-guards.

Reviewed By: dneiter

Differential Revision: D21701324

fbshipit-source-id: 0228f8fca880983a596b941d1afff025127f70a8